+++ bootique-linkmove-rest-cayenne41/src/main/java/io/bootique/linkmove/rest/cayenne41/LinkMoveRestModule.java
@@ -20,5 +20,8 @@ package io.bootique.linkmove.rest.cayenne41;
 
 import io.bootique.ConfigModule;
 
+/**
+ * @since 1.1
+ */
 public class LinkMoveRestModule extends ConfigModule {
 }