+++ bootique-linkmove-rest-cayenne41/src/main/java/io/bootique/linkmove/rest/cayenne41/LinkMoveRestModuleProvider.java
@@ -21,6 +21,9 @@ package io.bootique.linkmove.rest.cayenne41;
 import com.google.inject.Module;
 import io.bootique.BQModuleProvider;
 
+/**
+ * @since 1.1
+ */
 public class LinkMoveRestModuleProvider implements BQModuleProvider {
 
     @Override