+++ bootique-linkmove-rest/src/main/java/io/bootique/linkmove/rest/LinkMoveRestModule.java
@@ -20,7 +20,8 @@ package io.bootique.linkmove.rest;
 
 import io.bootique.ConfigModule;
 
+/**
+ * @since 1.1
+ */
 public class LinkMoveRestModule extends ConfigModule {
-
-
 }